Split API models into feature modules
