+++ apps/server/README.md
@@ -26,7 +26,7 @@ Backend ownership boundaries:
 - `infra/config/`: runtime settings stores used by recording and runtime services.
 - `use_cases/diagnostics/`: post-stop analysis/findings logic; `run_data_preparation.py` owns `PreparedRunData` plus timing/speed/phase/sensor preparation, `order_analysis.py` now keeps the core matching/scoring/assembly primitives, `order_pipeline.py` owns the order-finding orchestration/session flow above those primitives, `order_heuristics.py` owns the pure scoring/filter heuristics, `peak_table.py` owns peak-table row ranking for persisted/report payloads, `plots.py` keeps plot-series plus FFT/spectrogram orchestration, and the package-level API still re-exports shared vehicle-order helpers used by live telemetry.
 - `use_cases/run/`: recording orchestration; `lifecycle_state.py` owns the in-memory run session state, `persistence_writer.py` owns history-write coordination and retry bookkeeping, `logger.py` owns the coordinator, and `post_analysis.py` owns the background analysis queue above the injected persistence/analysis/error boundary.
-- `adapters/persistence/`, `shared/`, and `use_cases/history/`: SQLite persistence, shared typed contracts/pure helpers (including run-log decoding, the `shared/ports.py` module for `RunPersistence`, `ClientTracker`, `SignalSource`, `SpeedProvider`, and `SettingsReader`, and vehicle-order math), car library loading, and history/report/export services. `shared/types/` now stays focused on payload/model types, while `use_cases/history/report_interpretation.py` owns pure report-domain interpretation used by PDF mapping and `adapters/persistence/history_db/` keeps `HistoryDB` as the public facade above internal `_run_lifecycle.py`, `_sample_io.py`, and `_queries.py` sections plus shared `_schema.py` and `_samples.py` helpers.
+- `adapters/persistence/`, `shared/`, and `use_cases/history/`: SQLite persistence, shared typed contracts/pure helpers (including run-log decoding, the `shared/ports.py` module for `RunPersistence`, `ClientTracker`, `SignalSource`, `SpeedProvider`, and `SettingsReader`, and vehicle-order math), car library loading, and history/report/export services. `shared/types/` now stays focused on payload/model types, with feature-scoped Pydantic HTTP contracts in `shared/types/api_models/`, while `use_cases/history/report_interpretation.py` owns pure report-domain interpretation used by PDF mapping and `adapters/persistence/history_db/` keeps `HistoryDB` as the public facade above internal `_run_lifecycle.py`, `_sample_io.py`, and `_queries.py` sections plus shared `_schema.py` and `_samples.py` helpers.
 - `adapters/pdf/`: PDF/report rendering pipeline.
 - `use_cases/updates/`: wheel-based update flow.
 - `adapters/hotspot/`: Wi-Fi AP monitoring, parsing, and self-heal infrastructure.
