--- apps/server/vibesensor/shared/types/api_models.py
@@ -1,781 +1,0 @@
-"""Pydantic request/response models for the VibeSensor HTTP API.
-
-Separated from ``api.py`` to keep routing logic distinct from data contracts.
-"""
-
-from __future__ import annotations
-
-from typing import Annotated, Literal, TypeAlias
-
-from pydantic import BaseModel, ConfigDict, Field, model_validator
-
-from vibesensor.shared.types.backend_types import (
-    LanguageCode,
-    ResolvedSpeedSource,
-    SpeedSourceKind,
-    SpeedUnitCode,
-)
-from vibesensor.shared.types.payload_types import ClientApiRow
-
-__all__ = [
-    # Request models
-    "IdentifyRequest",
-    "SetLocationRequest",
-    "AnalysisSettingsRequest",
-    "LanguageRequest",
-    "SpeedUnitRequest",
-    "CarUpsertRequest",
-    "ActiveCarRequest",
-    "SpeedSourceRequest",
-    "UpdateStartRequest",
-    "EspFlashStartRequest",
-    "SensorRequest",
-    # Response models
-    "HealthDataLossResponse",
-    "HealthIntakeStatsResponse",
-    "HealthPersistenceResponse",
-    "HealthResponse",
-    "CarResponse",
-    "CarsResponse",
-    "SpeedSourceResponse",
-    "SpeedSourceStatusResponse",
-    "SensorConfigResponse",
-    "SensorsResponse",
-    "LanguageResponse",
-    "SpeedUnitResponse",
-    "ClientsResponse",
-    "LocationOptionResponse",
-    "ClientLocationsResponse",
-    "AnalysisSettingsResponse",
-    "IdentifyResponse",
-    "SetClientLocationResponse",
-    "RemoveClientResponse",
-    "RecordingStatusResponse",
-    "HistoryListEntryResponse",
-    "HistoryListResponse",
-    "HistoryRunResponse",
-    "HistoryInsightWarningResponse",
-    "AmplitudeMetric",
-    "FindingEvidenceMetrics",
-    "FindingPayload",
-    "HistoryInsightsResponse",
-    "LocationHotspotPayload",
-    "MatchedPoint",
-    "PhaseEvidence",
-    "DeleteHistoryRunResponse",
-    "UpdateRuntimeResponse",
-    "UpdateIssueResponse",
-    "UpdateStatusResponse",
-    "UpdateStartResponse",
-    "UpdateCancelResponse",
-    "EspSerialPortResponse",
-    "EspFlashPortsResponse",
-    "EspFlashStatusResponse",
-    "EspFlashStartResponse",
-    "EspFlashCancelResponse",
-    "EspFlashLogsResponse",
-    "EspFlashHistoryEntryResponse",
-    "EspFlashHistoryResponse",
-    "CarLibraryBrandsResponse",
-    "CarLibraryTypesResponse",
-    "CarLibraryGearboxEntry",
-    "CarLibraryTireOptionEntry",
-    "CarLibraryVariantEntry",
-    "CarLibraryModelEntry",
-    "CarLibraryModelsResponse",
-]
-
-ApiPayloadObject: TypeAlias = dict[str, object]
-ApiPayloadValue: TypeAlias = None | bool | int | float | str | list[object] | ApiPayloadObject
-
-# ---------------------------------------------------------------------------
-# Shared base classes
-# ---------------------------------------------------------------------------
-
-
-class _FrozenBase(BaseModel):
-    """Immutable base for request models (constructed once, never mutated)."""
-
-    model_config = ConfigDict(frozen=True)
-
-
-class _ExtraAllowBase(BaseModel):
-    """Base for models that accept arbitrary extra fields."""
-
-    model_config = ConfigDict(extra="allow")
-
-
-# ---------------------------------------------------------------------------
-# Request models
-# ---------------------------------------------------------------------------
-
-
-class IdentifyRequest(_FrozenBase):
-    """Request body for the ``/api/clients/{id}/identify`` endpoint."""
-
-    duration_ms: int = Field(default=1500, ge=100, le=60_000)
-
-
-class SetLocationRequest(_FrozenBase):
-    """Request body for setting the sensor location code."""
-
-    location_code: str = Field(min_length=0, max_length=64)
-
-
-class AnalysisSettingsRequest(_FrozenBase):
-    """Request body for updating vehicle analysis settings (tire geometry, gear ratios, etc.)."""
-
-    tire_width_mm: float | None = Field(default=None, gt=0)
-    tire_aspect_pct: float | None = Field(default=None, gt=0)
-    rim_in: float | None = Field(default=None, gt=0)
-    final_drive_ratio: float | None = Field(default=None, gt=0)
-    current_gear_ratio: float | None = Field(default=None, gt=0)
-    wheel_bandwidth_pct: float | None = Field(default=None, gt=0)
-    driveshaft_bandwidth_pct: float | None = Field(default=None, gt=0)
-    engine_bandwidth_pct: float | None = Field(default=None, gt=0)
-    speed_uncertainty_pct: float | None = Field(default=None, ge=0)
-    tire_diameter_uncertainty_pct: float | None = Field(default=None, ge=0)
-    final_drive_uncertainty_pct: float | None = Field(default=None, ge=0)
-    gear_uncertainty_pct: float | None = Field(default=None, ge=0)
-    min_abs_band_hz: float | None = Field(default=None, ge=0)
-    max_band_half_width_pct: float | None = Field(default=None, gt=0)
-    tire_deflection_factor: float | None = Field(default=None, ge=0.85, le=1.0)
-
-
-class LanguageRequest(_FrozenBase):
-    """Request body for changing the UI language."""
-
-    language: LanguageCode
-
-
-class SpeedUnitRequest(_FrozenBase):
-    """Request body for changing the displayed speed unit."""
-
-    speedUnit: SpeedUnitCode
-
-
-class CarUpsertRequest(_FrozenBase):
-    """Request body for creating or updating a car profile."""
-
-    name: Annotated[str, Field(min_length=1, max_length=64)] | None = None
-    type: Annotated[str, Field(min_length=1, max_length=64)] | None = None
-    aspects: dict[str, float] | None = None
-    variant: Annotated[str, Field(min_length=1, max_length=64)] | None = None
-
-
-class ActiveCarRequest(_FrozenBase):
-    """Request body for selecting the active car profile."""
-
-    carId: str = Field(min_length=1)
-
-
-class SpeedSourceRequest(_FrozenBase):
-    """Request body for configuring the speed source (GPS, manual, OBD2, etc.)."""
-
-    speedSource: SpeedSourceKind | None = None
-    manualSpeedKph: float | None = Field(default=None, ge=0, le=500)
-    staleTimeoutS: float | None = Field(default=None, ge=3, le=120)
-
-
-class UpdateStartRequest(_FrozenBase):
-    """Request body to start an OTA software update (provides Wi-Fi credentials)."""
-
-    ssid: str = Field(min_length=1, max_length=64)
-    password: str = Field(default="", max_length=128)
-
-
-class EspFlashStartRequest(_FrozenBase):
-    """Request body to start an ESP32 firmware flash job."""
-
-    port: str | None = None
-    auto_detect: bool = True
-
-    @model_validator(mode="after")
-    def _require_port_when_not_auto_detect(self) -> EspFlashStartRequest:
-        """When auto_detect is False a port must be provided explicitly."""
-        if not self.auto_detect and not self.port:
-            raise ValueError("port is required when auto_detect is False")
-        return self
-
-
-class SensorRequest(_FrozenBase):
-    """Request body for updating sensor name and location."""
-
-    name: str | None = Field(default=None, min_length=1, max_length=64)
-    location_code: str | None = Field(default=None, max_length=64)
-
-
-# ---------------------------------------------------------------------------
-# Response models
-# ---------------------------------------------------------------------------
-
-
-class HealthDataLossResponse(BaseModel):
-    """Response body for aggregated client data-loss counters."""
-
-    tracked_clients: int
-    affected_clients: int
-    frames_dropped: int
-    queue_overflow_drops: int
-    server_queue_drops: int
-    parse_errors: int
-
-
-class HealthPersistenceResponse(BaseModel):
-    """Response body for persistence health details."""
-
-    write_error: str | None
-    analysis_in_progress: bool
-    analysis_queue_depth: int = 0
-    analysis_queue_max_depth: int = 0
-    analysis_active_run_id: str | None = None
-    analysis_started_at: float | None = None
-    analysis_elapsed_s: float | None = None
-    analysis_queue_oldest_age_s: float | None = None
-    analyzing_run_count: int = 0
-    analyzing_oldest_age_s: float | None = None
-    samples_written: int = 0
-    samples_dropped: int = 0
-    last_completed_run_id: str | None = None
-    last_completed_run_error: str | None = None
-
-
-class HealthIntakeStatsResponse(BaseModel):
-    """Response body for processing intake timing and throughput counters."""
-
-    total_ingested_samples: int
-    total_compute_calls: int
-    last_compute_duration_s: float
-    last_compute_all_duration_s: float
-    last_ingest_duration_s: float
-
-
-class HealthResponse(BaseModel):
-    """Response body for the server health check endpoint."""
-
-    status: Literal["ok", "warn", "degraded"]
-    startup_state: str
-    startup_phase: str
-    startup_error: str | None
-    startup_warnings: list[str] = []
-    background_task_failures: dict[str, str]
-    processing_state: str
-    processing_failures: int
-    processing_failure_categories: dict[str, int]
-    processing_last_failure: str | None
-    sample_rate_mismatch_count: int
-    frame_size_mismatch_count: int
-
-    degradation_reasons: list[str]
-    data_loss: HealthDataLossResponse
-    persistence: HealthPersistenceResponse
-    intake_stats: HealthIntakeStatsResponse
-
-    tick_duration_s: float = 0.0
-    max_tick_duration_s: float = 0.0
-    tick_count: int = 0
-    db_last_write_duration_s: float = 0.0
-    db_max_write_duration_s: float = 0.0
-
-
-class CarResponse(BaseModel):
-    """Response body representing a single car profile."""
-
-    id: str
-    name: str
-    type: str
-    aspects: dict[str, float]
-    variant: str | None = None
-
-
-class CarsResponse(BaseModel):
-    """Response body for the list of all car profiles with the active car ID."""
-
-    cars: list[CarResponse]
-    activeCarId: str | None
-
-
-class SpeedSourceResponse(BaseModel):
-    """Response body for the current speed-source configuration."""
-
-    speedSource: SpeedSourceKind
-    manualSpeedKph: float | None
-    staleTimeoutS: float
-
-
-class SpeedSourceStatusResponse(BaseModel):
-    """Response body for the live GPS/speed-source connection status."""
-
-    gps_enabled: bool
-    connection_state: str
-    device: str | None
-    fix_mode: int | None
-    fix_dimension: Literal["3d", "2d", "none"]
-    speed_confidence: Literal["low", "medium", "high"]
-    epx_m: float | None
-    epy_m: float | None
-    epv_m: float | None
-    last_update_age_s: float | None
-    raw_speed_kmh: float | None
-    effective_speed_kmh: float | None
-    last_error: str | None
-    reconnect_delay_s: float | None
-    fallback_active: bool
-    speed_source: ResolvedSpeedSource
-    stale_timeout_s: float
-
-
-class SensorConfigResponse(BaseModel):
-    """Response body with persisted config for a single sensor (name, location_code)."""
-
-    name: str
-    location_code: str
-
-
-class SensorsResponse(BaseModel):
-    """Response body mapping MAC addresses to sensor config responses."""
-
-    sensorsByMac: dict[str, SensorConfigResponse]
-
-
-class LanguageResponse(BaseModel):
-    """Response body confirming the active UI language."""
-
-    language: str
-
-
-class SpeedUnitResponse(BaseModel):
-    """Response body confirming the active speed unit."""
-
-    speedUnit: str
-
-
-class ClientsResponse(BaseModel):
-    """Response body listing all currently-connected sensor clients."""
-
-    clients: list[ClientApiRow]
-
-
-class LocationOptionResponse(BaseModel):
-    """A single sensor-location option (code + human-readable label)."""
-
-    code: str
-    label: str
-
-
-class ClientLocationsResponse(BaseModel):
-    """Response body with available sensor-location options."""
-
-    locations: list[LocationOptionResponse]
-
-
-class AnalysisSettingsResponse(BaseModel):
-    """Response body reflecting the current validated analysis settings."""
-
-    tire_width_mm: float
-    tire_aspect_pct: float
-    rim_in: float
-    final_drive_ratio: float
-    current_gear_ratio: float
-    wheel_bandwidth_pct: float
-    driveshaft_bandwidth_pct: float
-    engine_bandwidth_pct: float
-    speed_uncertainty_pct: float
-    tire_diameter_uncertainty_pct: float
-    final_drive_uncertainty_pct: float
-    gear_uncertainty_pct: float
-    min_abs_band_hz: float
-    max_band_half_width_pct: float
-    tire_deflection_factor: float
-
-
-class IdentifyResponse(BaseModel):
-    """Response body for a sensor identify (blink) command."""
-
-    status: str
-    cmd_seq: int | None = None
-
-
-class SetClientLocationResponse(BaseModel):
-    """Response body confirming the new location assignment for a client."""
-
-    id: str
-    mac_address: str
-    location_code: str
-    name: str
-
-
-class RemoveClientResponse(BaseModel):
-    """Response body confirming removal of a disconnected client."""
-
-    id: str
-    status: str
-
-
-class RecordingStatusResponse(BaseModel):
-    """Response body with the current recording (run-logging) status."""
-
-    enabled: bool
-    current_file: str | None
-    run_id: str | None
-    write_error: str | None
-    analysis_in_progress: bool
-    samples_written: int = 0
-    samples_dropped: int = 0
-    last_completed_run_id: str | None = None
-    last_completed_run_error: str | None = None
-
-
-class HistoryListEntryResponse(BaseModel):
-    """Response body for a single history-run list row."""
-
-    run_id: str
-    status: str
-    start_time_utc: str
-    end_time_utc: str | None = None
-    created_at: str
-    sample_count: int
-    error_message: str | None = None
-
-
-class HistoryListResponse(BaseModel):
-    """Response body listing recorded run summaries."""
-
-    runs: list[HistoryListEntryResponse]
-
-
-class HistoryRunResponse(_ExtraAllowBase):
-    """Response body for a single history run with metadata and optional analysis."""
-
-    run_id: str
-    status: str
-    metadata: ApiPayloadObject = Field(default_factory=dict)
-    analysis: ApiPayloadObject | None = None
-
-
-class HistoryInsightWarningResponse(BaseModel):
-    """Response body for a localized history/run trust warning."""
-
-    code: str
-    severity: Literal["warn", "error"]
-    applies_to: str
-    title: str
-    detail: str | None = None
-
-
-class MatchedPoint(_ExtraAllowBase):
-    """HTTP contract for one serialized finding matched-point observation."""
-
-    t_s: float | None = None
-    speed_kmh: float | None = None
-    predicted_hz: float | None = None
-    matched_hz: float | None = None
-    rel_error: float | None = None
-    amp: float | None = None
-    location: str | None = None
-    phase: str | None = None
-
-
-class PhaseEvidence(_ExtraAllowBase):
-    """HTTP contract for optional driving-phase evidence attached to a finding."""
-
-    cruise_fraction: float | None = None
-    phases_detected: list[str] = Field(default_factory=list)
-
-
-class AmplitudeMetric(_ExtraAllowBase):
-    """HTTP contract for finding amplitude/strength metadata.
-
-    The runtime payload currently emits both a compact
-    ``{"vibration_strength_db": ...}`` shape and richer
-    ``{"name", "value", "units", "definition"}`` objects in tests/report helpers,
-    so the schema keeps both representations legal without changing runtime behavior.
-    """
-
-    name: str | None = None
-    value: float | None = None
-    units: str | None = None
-    definition: ApiPayloadValue = None
-    vibration_strength_db: float | None = None
-
-
-class LocationHotspotPayload(_ExtraAllowBase):
-    """HTTP contract for serialized location-hotspot evidence."""
-
-    dominance_ratio: float | None = None
-    location_count: int | None = None
-    top_location: str | None = None
-    second_location: str | None = None
-    ambiguous_location: bool | None = None
-    ambiguous_locations: list[str] = Field(default_factory=list)
-    localization_confidence: float | None = None
-    weak_spatial_separation: bool | None = None
-
-
-class FindingEvidenceMetrics(_ExtraAllowBase):
-    """HTTP contract for serialized evidence metrics attached to a finding."""
-
-    match_rate: float | None = None
-    global_match_rate: float | None = None
-    focused_speed_band: str | None = None
-    mean_relative_error: float | None = None
-    mean_noise_floor_db: float | None = None
-    vibration_strength_db: float | None = None
-    possible_samples: int | None = None
-    matched_samples: int | None = None
-    frequency_correlation: float | None = None
-    per_phase_confidence: dict[str, float] | None = None
-    phases_with_evidence: int | None = None
-    presence_ratio: float | None = None
-    median_intensity_db: float | None = None
-    p95_intensity_db: float | None = None
-    max_intensity_db: float | None = None
-    burstiness: float | None = None
-    run_noise_baseline_db: float | None = None
-    median_relative_to_run_noise: float | None = None
-    p95_relative_to_run_noise: float | None = None
-    sample_count: int | None = None
-    total_samples: int | None = None
-    spatial_concentration: float | None = None
-    spatial_uniformity: float | None = None
-    speed_uniformity: float | None = None
-
-
-class FindingPayload(_ExtraAllowBase):
-    """HTTP contract for one serialized finding in analysis history payloads.
-
-    This schema mirrors ``shared.boundaries.analysis_payload.FindingPayload``.
-    It intentionally includes a few presentation-oriented projections
-    (``evidence_summary``, ``frequency_hz_or_order``, ``amplitude_metric``,
-    and the confidence label fields) alongside the domain-owned finding data.
-    """
-
-    finding_id: str
-    finding_key: str | None = None
-    suspected_source: str
-    evidence_summary: ApiPayloadValue
-    frequency_hz_or_order: ApiPayloadValue
-    amplitude_metric: AmplitudeMetric
-    confidence: float | None
-    finding_kind: str | None = None
-    severity: str | None = None
-    confidence_label_key: str | None = None
-    confidence_tone: str | None = None
-    confidence_pct: str | None = None
-    matched_points: list[MatchedPoint] = Field(default_factory=list)
-    location_hotspot: LocationHotspotPayload | None = None
-    strongest_location: str | None = None
-    strongest_speed_band: str | None = None
-    dominant_phase: str | None = None
-    dominance_ratio: float | None = None
-    weak_spatial_separation: bool | None = None
-    diffuse_excitation: bool | None = None
-    phase_evidence: PhaseEvidence | None = None
-    evidence_metrics: FindingEvidenceMetrics | None = None
-    ranking_score: float | None = None
-    peak_classification: str | None = None
-    signatures_observed: list[str] = Field(default_factory=list)
-    order: str | None = None
-
-
-class HistoryInsightsResponse(_ExtraAllowBase):
-    """Response body with aggregated diagnostic insights for a run."""
-
-    run_id: str | None = None
-    status: str | None = None
-    warnings: list[HistoryInsightWarningResponse] = Field(default_factory=list)
-    findings: list[FindingPayload] = Field(default_factory=list)
-    top_causes: list[FindingPayload] = Field(default_factory=list)
-
-
-class DeleteHistoryRunResponse(BaseModel):
-    """Response body confirming deletion of a history run."""
-
-    run_id: str
-    status: str
-
-
-class UpdateIssueResponse(BaseModel):
-    """Response body for a single issue raised during an OTA update phase."""
-
-    phase: str
-    message: str
-    detail: str
-
-
-class UpdateRuntimeResponse(BaseModel):
-    """Response body for updater runtime/build verification details."""
-
-    version: str
-    commit: str
-    ui_source_hash: str
-    static_assets_hash: str
-    static_build_source_hash: str
-    static_build_commit: str
-    assets_verified: bool
-    has_packaged_static: bool
-
-
-class UpdateStatusResponse(BaseModel):
-    """Response body for the full OTA update job status."""
-
-    state: str
-    phase: str
-    started_at: float | None = None
-    finished_at: float | None = None
-    last_success_at: float | None = None
-    phase_started_at: float | None = None
-    phase_elapsed_s: float | None = None
-    updated_at: float | None = None
-    ssid: str
-    issues: list[UpdateIssueResponse]
-    log_tail: list[str]
-    exit_code: int | None = None
-    runtime: UpdateRuntimeResponse
-
-
-class UpdateStartResponse(BaseModel):
-    """Response body confirming that an OTA update job has started."""
-
-    status: str
-    ssid: str
-
-
-class UpdateCancelResponse(BaseModel):
-    """Response body confirming whether an OTA update job was cancelled."""
-
-    cancelled: bool
-
-
-class EspSerialPortResponse(BaseModel):
-    """Response body describing a single detected serial port."""
-
-    port: str
-    description: str
-    vid: int | None = None
-    pid: int | None = None
-    serial_number: str | None = None
-
-
-class EspFlashPortsResponse(BaseModel):
-    """Response body listing detected serial ports for ESP32 flashing."""
-
-    ports: list[EspSerialPortResponse]
-
-
-class EspFlashStatusResponse(BaseModel):
-    """Response body for the current ESP32 flash job status."""
-
-    state: str
-    phase: str
-    job_id: int | None = None
-    selected_port: str | None = None
-    auto_detect: bool
-    started_at: float | None = None
-    finished_at: float | None = None
-    last_success_at: float | None = None
-    exit_code: int | None = None
-    error: str | None = None
-    log_count: int
-
-
-class EspFlashStartResponse(BaseModel):
-    """Response body confirming that an ESP32 flash job has been queued."""
-
-    status: str
-    job_id: int
-
-
-class EspFlashCancelResponse(BaseModel):
-    """Response body confirming whether an ESP32 flash job was cancelled."""
-
-    cancelled: bool
-
-
-class EspFlashLogsResponse(BaseModel):
-    """Response body with a page of ESP32 flash log lines."""
-
-    from_index: int
-    next_index: int
-    lines: list[str]
-
-
-class EspFlashHistoryEntryResponse(BaseModel):
-    """Response body for a single historical ESP32 flash job."""
-
-    job_id: int
-    state: str
-    selected_port: str | None = None
-    auto_detect: bool
-    started_at: float
-    finished_at: float | None = None
-    exit_code: int | None = None
-    error: str | None = None
-
-
-class EspFlashHistoryResponse(BaseModel):
-    """Response body listing all past ESP32 flash job attempts."""
-
-    attempts: list[EspFlashHistoryEntryResponse]
-
-
-class CarLibraryBrandsResponse(BaseModel):
-    """Response body listing available car manufacturer brands."""
-
-    brands: list[str]
-
-
-class CarLibraryTypesResponse(BaseModel):
-    """Response body listing available car body types."""
-
-    types: list[str]
-
-
-class CarLibraryGearboxEntry(_ExtraAllowBase):
-    """A gearbox option from the car library (gear ratios)."""
-
-    name: str = Field(min_length=1)
-    final_drive_ratio: float = Field(gt=0)
-    top_gear_ratio: float = Field(gt=0)
-
-
-class CarLibraryTireOptionEntry(_ExtraAllowBase):
-    """A tire size option from the car library."""
-
-    name: str = Field(min_length=1)
-    tire_width_mm: float = Field(gt=0)
-    tire_aspect_pct: float = Field(gt=0)
-    rim_in: float = Field(gt=0)
-
-
-class CarLibraryVariantEntry(_ExtraAllowBase):
-    """A specific variant/trim of a car library model entry."""
-
-    name: str = Field(min_length=1)
-    engine: str | None = None
-    drivetrain: Literal["FWD", "RWD", "AWD"]
-    gearboxes: list[CarLibraryGearboxEntry] | None = None
-    tire_options: list[CarLibraryTireOptionEntry] | None = None
-    tire_width_mm: float | None = Field(default=None, gt=0)
-    tire_aspect_pct: float | None = Field(default=None, gt=0)
-    rim_in: float | None = Field(default=None, gt=0)
-
-
-class CarLibraryModelEntry(_ExtraAllowBase):
-    """A full car library entry with brand, model, tire options, and variants."""
-
-    brand: str
-    type: str
-    model: str
-    gearboxes: list[CarLibraryGearboxEntry] = Field(min_length=1)
-    tire_options: list[CarLibraryTireOptionEntry] = Field(min_length=1)
-    tire_width_mm: float = Field(gt=0)
-    tire_aspect_pct: float = Field(gt=0)
-    rim_in: float = Field(gt=0)
-    variants: list[CarLibraryVariantEntry] = Field(default_factory=list)
-
-
-class CarLibraryModelsResponse(BaseModel):
-    """Response body listing car library model entries."""
-
-    models: list[CarLibraryModelEntry]
