+++ apps/server/vibesensor/shared/types/api_models/__init__.py
@@ -1,0 +1,152 @@
+"""Feature-scoped Pydantic request/response models for the VibeSensor HTTP API."""
+
+from __future__ import annotations
+
+from .base import ApiPayloadObject, ApiPayloadValue, _ExtraAllowBase, _FrozenBase
+from .car_library import (
+    CarLibraryBrandsResponse,
+    CarLibraryGearboxEntry,
+    CarLibraryModelEntry,
+    CarLibraryModelsResponse,
+    CarLibraryTireOptionEntry,
+    CarLibraryTypesResponse,
+    CarLibraryVariantEntry,
+)
+from .clients import (
+    ClientLocationsResponse,
+    ClientsResponse,
+    IdentifyRequest,
+    IdentifyResponse,
+    LocationOptionResponse,
+    RemoveClientResponse,
+    SetClientLocationResponse,
+    SetLocationRequest,
+)
+from .health import (
+    HealthDataLossResponse,
+    HealthIntakeStatsResponse,
+    HealthPersistenceResponse,
+    HealthResponse,
+)
+from .history import (
+    AmplitudeMetric,
+    DeleteHistoryRunResponse,
+    FindingEvidenceMetrics,
+    FindingPayload,
+    HistoryInsightsResponse,
+    HistoryInsightWarningResponse,
+    HistoryListEntryResponse,
+    HistoryListResponse,
+    HistoryRunResponse,
+    LocationHotspotPayload,
+    MatchedPoint,
+    PhaseEvidence,
+)
+from .recording import RecordingStatusResponse
+from .settings import (
+    ActiveCarRequest,
+    AnalysisSettingsRequest,
+    AnalysisSettingsResponse,
+    CarResponse,
+    CarsResponse,
+    CarUpsertRequest,
+    LanguageRequest,
+    LanguageResponse,
+    SensorConfigResponse,
+    SensorRequest,
+    SensorsResponse,
+    SpeedSourceRequest,
+    SpeedSourceResponse,
+    SpeedSourceStatusResponse,
+    SpeedUnitRequest,
+    SpeedUnitResponse,
+)
+from .updates import (
+    EspFlashCancelResponse,
+    EspFlashHistoryEntryResponse,
+    EspFlashHistoryResponse,
+    EspFlashLogsResponse,
+    EspFlashPortsResponse,
+    EspFlashStartRequest,
+    EspFlashStartResponse,
+    EspFlashStatusResponse,
+    EspSerialPortResponse,
+    UpdateCancelResponse,
+    UpdateIssueResponse,
+    UpdateRuntimeResponse,
+    UpdateStartRequest,
+    UpdateStartResponse,
+    UpdateStatusResponse,
+)
+
+__all__ = [
+    # Request models
+    "IdentifyRequest",
+    "SetLocationRequest",
+    "AnalysisSettingsRequest",
+    "LanguageRequest",
+    "SpeedUnitRequest",
+    "CarUpsertRequest",
+    "ActiveCarRequest",
+    "SpeedSourceRequest",
+    "UpdateStartRequest",
+    "EspFlashStartRequest",
+    "SensorRequest",
+    # Response models
+    "HealthDataLossResponse",
+    "HealthIntakeStatsResponse",
+    "HealthPersistenceResponse",
+    "HealthResponse",
+    "CarResponse",
+    "CarsResponse",
+    "SpeedSourceResponse",
+    "SpeedSourceStatusResponse",
+    "SensorConfigResponse",
+    "SensorsResponse",
+    "LanguageResponse",
+    "SpeedUnitResponse",
+    "ClientsResponse",
+    "LocationOptionResponse",
+    "ClientLocationsResponse",
+    "AnalysisSettingsResponse",
+    "IdentifyResponse",
+    "SetClientLocationResponse",
+    "RemoveClientResponse",
+    "RecordingStatusResponse",
+    "HistoryListEntryResponse",
+    "HistoryListResponse",
+    "HistoryRunResponse",
+    "HistoryInsightWarningResponse",
+    "AmplitudeMetric",
+    "FindingEvidenceMetrics",
+    "FindingPayload",
+    "HistoryInsightsResponse",
+    "LocationHotspotPayload",
+    "MatchedPoint",
+    "PhaseEvidence",
+    "DeleteHistoryRunResponse",
+    "UpdateRuntimeResponse",
+    "UpdateIssueResponse",
+    "UpdateStatusResponse",
+    "UpdateStartResponse",
+    "UpdateCancelResponse",
+    "EspSerialPortResponse",
+    "EspFlashPortsResponse",
+    "EspFlashStatusResponse",
+    "EspFlashStartResponse",
+    "EspFlashCancelResponse",
+    "EspFlashLogsResponse",
+    "EspFlashHistoryEntryResponse",
+    "EspFlashHistoryResponse",
+    "CarLibraryBrandsResponse",
+    "CarLibraryTypesResponse",
+    "CarLibraryGearboxEntry",
+    "CarLibraryTireOptionEntry",
+    "CarLibraryVariantEntry",
+    "CarLibraryModelEntry",
+    "CarLibraryModelsResponse",
+    "ApiPayloadObject",
+    "ApiPayloadValue",
+    "_FrozenBase",
+    "_ExtraAllowBase",
+]
