+++ apps/server/vibesensor/shared/types/api_models/settings.py
@@ -1,0 +1,170 @@
+"""Settings and analysis-configuration HTTP API models."""
+
+from __future__ import annotations
+
+from typing import Annotated, Literal
+
+from pydantic import BaseModel, Field
+
+from vibesensor.shared.types.backend_types import (
+    LanguageCode,
+    ResolvedSpeedSource,
+    SpeedSourceKind,
+    SpeedUnitCode,
+)
+
+from .base import _FrozenBase
+
+
+class AnalysisSettingsRequest(_FrozenBase):
+    """Request body for updating vehicle analysis settings (tire geometry, gear ratios, etc.)."""
+
+    tire_width_mm: float | None = Field(default=None, gt=0)
+    tire_aspect_pct: float | None = Field(default=None, gt=0)
+    rim_in: float | None = Field(default=None, gt=0)
+    final_drive_ratio: float | None = Field(default=None, gt=0)
+    current_gear_ratio: float | None = Field(default=None, gt=0)
+    wheel_bandwidth_pct: float | None = Field(default=None, gt=0)
+    driveshaft_bandwidth_pct: float | None = Field(default=None, gt=0)
+    engine_bandwidth_pct: float | None = Field(default=None, gt=0)
+    speed_uncertainty_pct: float | None = Field(default=None, ge=0)
+    tire_diameter_uncertainty_pct: float | None = Field(default=None, ge=0)
+    final_drive_uncertainty_pct: float | None = Field(default=None, ge=0)
+    gear_uncertainty_pct: float | None = Field(default=None, ge=0)
+    min_abs_band_hz: float | None = Field(default=None, ge=0)
+    max_band_half_width_pct: float | None = Field(default=None, gt=0)
+    tire_deflection_factor: float | None = Field(default=None, ge=0.85, le=1.0)
+
+
+class LanguageRequest(_FrozenBase):
+    """Request body for changing the UI language."""
+
+    language: LanguageCode
+
+
+class SpeedUnitRequest(_FrozenBase):
+    """Request body for changing the displayed speed unit."""
+
+    speedUnit: SpeedUnitCode
+
+
+class CarUpsertRequest(_FrozenBase):
+    """Request body for creating or updating a car profile."""
+
+    name: Annotated[str, Field(min_length=1, max_length=64)] | None = None
+    type: Annotated[str, Field(min_length=1, max_length=64)] | None = None
+    aspects: dict[str, float] | None = None
+    variant: Annotated[str, Field(min_length=1, max_length=64)] | None = None
+
+
+class ActiveCarRequest(_FrozenBase):
+    """Request body for selecting the active car profile."""
+
+    carId: str = Field(min_length=1)
+
+
+class SpeedSourceRequest(_FrozenBase):
+    """Request body for configuring the speed source (GPS, manual, OBD2, etc.)."""
+
+    speedSource: SpeedSourceKind | None = None
+    manualSpeedKph: float | None = Field(default=None, ge=0, le=500)
+    staleTimeoutS: float | None = Field(default=None, ge=3, le=120)
+
+
+class SensorRequest(_FrozenBase):
+    """Request body for updating sensor name and location."""
+
+    name: str | None = Field(default=None, min_length=1, max_length=64)
+    location_code: str | None = Field(default=None, max_length=64)
+
+
+class CarResponse(BaseModel):
+    """Response body representing a single car profile."""
+
+    id: str
+    name: str
+    type: str
+    aspects: dict[str, float]
+    variant: str | None = None
+
+
+class CarsResponse(BaseModel):
+    """Response body for the list of all car profiles with the active car ID."""
+
+    cars: list[CarResponse]
+    activeCarId: str | None
+
+
+class SpeedSourceResponse(BaseModel):
+    """Response body for the current speed-source configuration."""
+
+    speedSource: SpeedSourceKind
+    manualSpeedKph: float | None
+    staleTimeoutS: float
+
+
+class SpeedSourceStatusResponse(BaseModel):
+    """Response body for the live GPS/speed-source connection status."""
+
+    gps_enabled: bool
+    connection_state: str
+    device: str | None
+    fix_mode: int | None
+    fix_dimension: Literal["3d", "2d", "none"]
+    speed_confidence: Literal["low", "medium", "high"]
+    epx_m: float | None
+    epy_m: float | None
+    epv_m: float | None
+    last_update_age_s: float | None
+    raw_speed_kmh: float | None
+    effective_speed_kmh: float | None
+    last_error: str | None
+    reconnect_delay_s: float | None
+    fallback_active: bool
+    speed_source: ResolvedSpeedSource
+    stale_timeout_s: float
+
+
+class SensorConfigResponse(BaseModel):
+    """Response body with persisted config for a single sensor (name, location_code)."""
+
+    name: str
+    location_code: str
+
+
+class SensorsResponse(BaseModel):
+    """Response body mapping MAC addresses to sensor config responses."""
+
+    sensorsByMac: dict[str, SensorConfigResponse]
+
+
+class LanguageResponse(BaseModel):
+    """Response body confirming the active UI language."""
+
+    language: str
+
+
+class SpeedUnitResponse(BaseModel):
+    """Response body confirming the active speed unit."""
+
+    speedUnit: str
+
+
+class AnalysisSettingsResponse(BaseModel):
+    """Response body reflecting the current validated analysis settings."""
+
+    tire_width_mm: float
+    tire_aspect_pct: float
+    rim_in: float
+    final_drive_ratio: float
+    current_gear_ratio: float
+    wheel_bandwidth_pct: float
+    driveshaft_bandwidth_pct: float
+    engine_bandwidth_pct: float
+    speed_uncertainty_pct: float
+    tire_diameter_uncertainty_pct: float
+    final_drive_uncertainty_pct: float
+    gear_uncertainty_pct: float
+    min_abs_band_hz: float
+    max_band_half_width_pct: float
+    tire_deflection_factor: float
